project/requirements.txt
stable-audio-tools
stable-audio-metrics
transformers>=4.45.0
torch
torchaudio
librosa
soundfile
numpy
pandas
pyyaml
tqdm
